The intention of this PR is to get an initial review of the JsonLab reader/writer APIs and see how user code changes during a port. These APIs are still being prototyped and hence I expect them to change. There are also some semantic & behavioral differences here, and I wanted to discover how this would impact customer code.

Please note that the implementation is not thoroughly tested and there are gaps in functionality (for instance escaping input while writing or proper validation and generating rich error messages while reading are still missing).

**Feedback welcome.**

**TODO:**
- JsonHubProtocol: I tried to go [as far as possible](https://github.com/aspnet/SignalR/commit/1418aee1d707496bac566d43a47b4de3e7f50a9e#diff-12cd032d012d9ed36a88025fbc95c54c) with the new APIs, but this relies on a lot more features, including a [JsonSerializer](https://github.com/aspnet/SignalR/commit/1418aee1d707496bac566d43a47b4de3e7f50a9e#diff-12cd032d012d9ed36a88025fbc95c54cR231) (for reading and [writing](https://github.com/aspnet/SignalR/commit/1418aee1d707496bac566d43a47b4de3e7f50a9e#diff-12cd032d012d9ed36a88025fbc95c54cR581) objects), use of [JToken](https://github.com/aspnet/SignalR/commit/1418aee1d707496bac566d43a47b4de3e7f50a9e#diff-12cd032d012d9ed36a88025fbc95c54cR190), and [Stream-based](https://github.com/aspnet/SignalR/commit/1418aee1d707496bac566d43a47b4de3e7f50a9e#diff-0e5591272fc4a9be570908d247813246R73) reader, none of which exist atm.

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT
  Job-QQAEKJ : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.2  
RunStrategy=Throughput  

```
**Before:**

|                Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|---------------------- |---------:|----------:|----------:|----------:|-------:|----------:|
| SuccessHandshakeAsync | 2.524 us | 0.0501 us | 0.0492 us | 396,118.5 | 0.0076 |     808 B |
|   ErrorHandshakeAsync | 2.560 us | 0.0527 us | 0.0440 us | 390,608.5 | 0.0076 |     864 B |

|                           Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|--------------------------------- |---------:|----------:|----------:|----------:|-------:|----------:|
| WriteResponse_MemoryBufferWriter | 1.568 us | 0.0130 us | 0.0121 us | 637,800.6 | 0.0038 |     464 B |

**After:**
1.3 - 1.7x faster

|                Method |     Mean |     Error |    StdDev |   Median |      Op/s |  Gen 0 | Allocated |
|---------------------- |---------:|----------:|----------:|---------:|----------:|-------:|----------:|
| SuccessHandshakeAsync | 1.956 us | 0.0184 us | 0.0172 us | 1.951 us | 511,313.2 | 0.0038 |     504 B |
|   ErrorHandshakeAsync | 1.470 us | 0.0293 us | 0.0456 us | 1.448 us | 680,161.4 | 0.0038 |     400 B |

2.8x faster

|                           Method |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
|--------------------------------- |---------:|---------:|---------:|------------:|-------:|----------:|
| WriteResponse_MemoryBufferWriter | 555.2 ns | 10.62 ns | 10.90 ns | 1,801,121.9 | 0.0010 |     144 B |

cc @glennc, @rynowak, @davidfowl, @joshfree