@ahsonkhan 2018-08-04
Slight improvement in HubConnectionContextBenchmark:

|                Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|---------------------- |---------:|----------:|----------:|----------:|-------:|----------:|
| SuccessHandshakeAsync | 1.803 us | 0.0292 us | 0.0273 us | 554,550.7 | 0.0038 |     504 B |
|   ErrorHandshakeAsync | 1.412 us | 0.0274 us | 0.0336 us | 708,118.3 | 0.0038 |     400 B |

New benchmarks (1.8x faster):

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT
  Job-DCZITF : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.2  
RunStrategy=Throughput  

```
|                                             Method |     Mean |     Error |    StdDev |         Op/s | Allocated |
|--------------------------------------------------- |---------:|----------:|----------:|-------------:|----------:|
|       ParsingHandshakeRequestMessage_ValidMessage1 | 28.87 ns | 0.6073 ns | 0.6993 ns | 34,640,926.1 |       0 B |
|       ParsingHandshakeRequestMessage_ValidMessage2 | 28.89 ns | 0.5950 ns | 0.6110 ns | 34,615,174.3 |       0 B |
|       ParsingHandshakeRequestMessage_ValidMessage3 | 28.93 ns | 0.4382 ns | 0.3659 ns | 34,563,747.4 |       0 B |
|        ParsingHandshakeRequestMessage_NotComplete1 | 29.21 ns | 0.5334 ns | 0.4728 ns | 34,240,370.2 |       0 B |
|     ParsingHandshakeResponseMessage_ValidMessages1 | 28.42 ns | 0.6852 ns | 0.7616 ns | 35,185,553.8 |       0 B |
|     ParsingHandshakeResponseMessage_ValidMessages2 | 28.90 ns | 0.3204 ns | 0.2675 ns | 34,598,520.1 |       0 B |
|     ParsingHandshakeResponseMessage_ValidMessages3 | 28.76 ns | 0.6023 ns | 0.8638 ns | 34,769,221.2 |       0 B |
|     ParsingHandshakeResponseMessage_ValidMessages4 | 28.04 ns | 0.2955 ns | 0.2764 ns | 35,666,853.5 |       0 B |
| ParsingHandshakeResponseMessage_GivesMinorVersion1 | 27.97 ns | 0.5031 ns | 0.4460 ns | 35,758,243.7 |       0 B |
| ParsingHandshakeResponseMessage_GivesMinorVersion2 | 28.04 ns | 0.4796 ns | 0.4487 ns | 35,658,047.1 |       0 B |

Using TextMessageParser.TryParseMessage adds quite a bit of overhead for the single segment case.