@davidfowl 2018-08-04
> Using TextMessageParser.TryParseMessage adds quite a bit of overhead for the single segment case.

Good to know, that's another advantage of using the JsonLab reader. It supports spans natively so you can do this trick easily. 

Looking forward to the replacing the entire JsonHubProtocol 😄.