@davidfowl 2018-08-05
> JsonHubProtocol: I tried to go as far as possible with the new APIs, but this relies on a lot more features, including a JsonSerializer (for reading and writing objects), use of JToken, and Stream-based reader, none of which exist atm.

The biggest feature here is to be able to serialize an arbitrary type into an .NET object. The reason `JToken` is used is because we're mostly trying to be reading in a forward only fashion if properties come in an order that we can avoid that.

For example, if we're parsing an incoming method invocation and we get the arguments before the method name, we need to store the entire payload then turn that into the right set of arguments:

Here's the method:

```C#
public class MyHub
{
    int Add(int a, int b) => a + b;
}
```

Here's the JSON to call it:

```JSON
{ "invocationId": "1", "target": "Add", "arguments": [1, 2]  }
```

In this case, we know the method name before we parse the arguments (in the JSON payload) so we can parse each array entry into the appropriate .NET type (in this case integers).

But imagine we got this payload:

```JSON
{ "arguments": [1, 2], "invocationId": "1", "target": "Add" }
```

The invocation still needs to work so we end up storing arguments as a JArray until we know what method what the argument types are (which we find out by looking at the "target" field)

https://github.com/aspnet/SignalR/blob/1418aee1d707496bac566d43a47b4de3e7f50a9e/src/Microsoft.AspNetCore.SignalR.Protocols.Json/Protocol/JsonHubProtocol.cs#L250-L274

This logic exists for places where the user controls the type. The "PayloadSerializer" needs to be able handle these user types.