@ahsonkhan 2018-08-07
2-2.5x faster and lower allocations.

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=2.1.302
  [Host]     : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT
  Job-LAVJEB : .NET Core 2.2.0-preview1-26618-02 (CoreCLR 4.6.26616.03, CoreFX 4.6.26617.01), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.2  
RunStrategy=Throughput  

```
**Before:**

|                                          Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|------------------------------------------------ |---------:|----------:|----------:|----------:|-------:|----------:|
| ParsingNegotiateResponseMessageSuccessForValid1 | 1.518 us | 0.0303 us | 0.0498 us | 658,635.4 | 0.0420 |   3.84 KB |
| ParsingNegotiateResponseMessageSuccessForValid2 | 1.195 us | 0.0227 us | 0.0190 us | 836,854.1 | 0.0439 |    3.6 KB |
| ParsingNegotiateResponseMessageSuccessForValid3 | 1.306 us | 0.0257 us | 0.0315 us | 765,430.6 | 0.0439 |   3.68 KB |
| ParsingNegotiateResponseMessageSuccessForValid4 | 2.024 us | 0.0405 us | 0.0554 us | 494,086.9 | 0.0458 |   4.09 KB |
| ParsingNegotiateResponseMessageSuccessForValid5 | 2.340 us | 0.0451 us | 0.0443 us | 427,396.1 | 0.0458 |    4.3 KB |

**After:**

|                                          Method |       Mean |     Error |    StdDev |        Op/s |  Gen 0 | Allocated |
|------------------------------------------------ |-----------:|----------:|----------:|------------:|-------:|----------:|
| ParsingNegotiateResponseMessageSuccessForValid1 |   535.7 ns | 11.153 ns | 15.996 ns | 1,866,642.1 | 0.0076 |     688 B |
| ParsingNegotiateResponseMessageSuccessForValid2 |   454.9 ns |  9.128 ns | 19.054 ns | 2,198,199.2 | 0.0072 |     664 B |
| ParsingNegotiateResponseMessageSuccessForValid3 |   570.5 ns |  7.812 ns |  6.523 ns | 1,752,747.3 | 0.0076 |     720 B |
| ParsingNegotiateResponseMessageSuccessForValid4 |   867.7 ns | 17.382 ns | 17.850 ns | 1,152,518.6 | 0.0095 |     888 B |
| ParsingNegotiateResponseMessageSuccessForValid5 | 1,146.7 ns | 22.925 ns | 38.303 ns |   872,069.0 | 0.0114 |    1152 B |