@ahsonkhan 2018-08-13
Does anyone else have any feedback around the API and its use within SignalR?