@BrennanConroy 2018-08-14
I think it looks good.