@anurse 2018-08-15
Haven't been able to dig deep but it looks pretty good. I have some naming comments on the writer API (it feels weird that it has the same structure as JSON.NET's API but slightly different names) but structurally I think it's fine.