@ahsonkhan 2018-08-15
> WriteAttribute is kind of weird. Attribute is a very XML word.
I think WriteNameValue fits better into JSON land.

I have received this feedback in the past, and I agree. I will change it.