@anurse 2018-08-15
Cool. Don't want to hijack this with a naming thread but it'll be good to cover that in an API review.

API looks good aside from some bikeshedding around naming 😛